Make compatible with SystemJS
Newer versions of Babel have discontinued support for CommonJS (eg: `exports`). This changes the exports to be ES6-compliant which must also take into account variable scoping on the client since variables declared on `window` are no longer considered to be global (but, `window` is). We also make a best effort to merge with existing objects.